Moved the peoplesearch folder underneath the "./src/modules" folder.
