+++ client/src/main.dev.ts
@@ -23,7 +23,7 @@
 
 import { bootstrap, module } from 'angular';
 import ConfigService from './services/peoplesearch-config.service.dev';
-import peopleSearchModule from './peoplesearch/peoplesearch.module';
+import peopleSearchModule from './modules/peoplesearch/peoplesearch.module';
 import PeopleService from './services/people.service.dev';
 import PwmService from './services/pwm.service.dev';
 import routes from './routes';
