+++ client/src/modules/peoplesearch/peoplesearch-cards.component.scss
