+++ client/src/modules/peoplesearch/peoplesearch.module.ts
@@ -30,9 +30,9 @@ import PeopleSearchTableComponent from './peoplesearch-table.component';
 import PeopleSearchCardsComponent from './peoplesearch-cards.component';
 import PersonCardDirective from './person-card.component';
 import PersonDetailsDialogComponent from './person-details-dialog.component';
-import LocalStorageService from '../services/local-storage.service';
-import PromiseService from '../services/promise.service';
-import uxModule from '../ux/ux.module';
+import LocalStorageService from '../../services/local-storage.service';
+import PromiseService from '../../services/promise.service';
+import uxModule from '../../ux/ux.module';
 
 require('./peoplesearch.scss');
