+++ client/src/modules/peoplesearch/peoplesearch.scss
@@ -78,7 +78,7 @@ body {
 }
 
 .ias-avatar {
-  background: transparent url('../../images/user.png') no-repeat center center;
+  background: transparent url('../../../images/user.png') no-repeat center center;
   background-size: contain;
 }
