+++ client/src/modules/peoplesearch/person-card.component.ts
@@ -22,10 +22,10 @@
 
 
 import {IAugmentedJQuery} from 'angular';
-import { IPerson } from '../models/person.model';
-import { IPeopleService } from '../services/people.service';
+import { IPerson } from '../../models/person.model';
+import { IPeopleService } from '../../services/people.service';
 
-const templateUrl = require('peoplesearch/person-card.component.html');
+const templateUrl = require('modules/peoplesearch/person-card.component.html');
 
 class PersonCardController {
     private details: any[]; // For large style cards
