+++ client/src/modules/peoplesearch/person-details-dialog.component.scss
